feat: 주문 취소 내역 생성 추가, 주문 취소 검증 로직 추가

+++ domain/domain-rdb/src/main/java/com/gamzabit/domain/user/UserAssetProcessor.java
@@ -26,7 +26,7 @@ public class UserAssetProcessor {
 
     public void withdrawTo(User user, AssetPrice withdrawAmount, String assetTypeString) {
         UserAssetEntity userAssetEntity =
-            userAssetQueryRepository.findByUserIdAndAssetName(user.id().longValue(), assetTypeString)
+            userAssetQueryRepository.findByUserIdAndAssetName(user.id(), assetTypeString)
                 .orElseThrow(() ->
                     new AssetNotFoundException("유저의 '" + assetTypeString + "' 자산이 존재하지 않습니다.", assetTypeString)
                 );
